pyproject.toml
[tool.poetry]
authors = ["t-ohtsuka89 <102138393+t-ohtsuka89@users.noreply.github.com>"]
description = ""
name = "gpuser"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
django = "^5.0"
python = "^3.12"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
paramiko = "^3.4.0"

[tool.poetry.group.dev.dependencies]
djlint = "^1.34.1"
pre-commit = "^3.6.0"
ruff = "^0.1.9"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
